Extract ViewModel creation to ProductCategoryViewModelBuilder type
